tensorrt/requirements.txt
onnx
onnxruntime
Pillow
numpy
